move user info to sub folder
