+++ frontend/src/components/rateMovie/RateMovieView.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import UserInfo from './rateSteps/UserInfo';
+import UserInfo from './userInfo/UserInfo';
 import Questions from './Questions/Questions';
 import Ratings from './ratings/ratings';
 
@@ -10,7 +10,7 @@ class RateMovieView extends Component {
             <div>
                 <div className="section loader">
                     <div className="title">rate-movie</div>
-                    {1 === 1 && <Ratings />}
+                    {1 === 1 && <UserInfo />}
                     {currentStep === 2 && <UserInfo />}
                     {currentStep === 3 && <UserInfo />}
                     {currentStep === 4 && <UserInfo />}
