+++ frontend/src/components/rateMovie/userInfo/UserInfoView.js
